telegram-bot/requirements.txt
python-telegram-bot>=21.0
supabase>=2.0.0
python-dotenv>=1.0.0
google-genai>=0.3.0
